mybatis changes

+++ src/main/java/by/tataranovich/leasingcompany/service/IContractService.java
@@ -5,5 +5,5 @@ import java.util.List;
 import by.tataranovich.leasingcompany.model.Contract;
 
 public interface IContractService extends DefaultService<Contract> {
-    List<Contract> getAllByLeasingCompanyId(int idLeasingCompany);
+    List<Contract> getAllByLeasingCompanyId(Long idLeasingCompany);
 }
